restructure component
